requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0 